moved game-general implementations to Game.cpp
