+++ snake/Game.h
@@ -1,8 +1,13 @@
-// A file containing general game functions and constants for terminal games
+// Contains general game functions and constants for terminal games
 // like snake and bmtron
-
 #ifndef GAME_H
 #define GAME_H
+
+#include <iostream>
+#include <cmath>
+#include <stdlib.h>
+#include <ncurses.h>
+#include <sys/ioctl.h>  // for get window size
 
 class GameVars {
 public:
@@ -22,16 +27,16 @@ struct Coords {
   char symbol;
 };
 
-// Print prompt
-void print_prompt();
-
-// Checks for user keyboard hit (req if getch is nonblocking)
-bool kbhit();
-
 // Init all the ncurses specific stuff
 void init_ncurses();
 
 // Sets the window size and board size, return true if changed
 bool set_game_size();
+
+// Checks for user keyboard hit (req if getch is nonblocking)
+bool kbhit();
+
+// Print prompt
+void print_prompt();
 
 #endif
