+++ snake/makefile
@@ -2,7 +2,7 @@ play: snake
 	./snake
 
 snake: snake_game.cpp
-	g++ snake_game.cpp -lncurses -O1 -o snake
+	g++ snake_game.cpp Game.cpp -lncurses -O1 -o snake
 
 clean:
 	\rm snake
