+++ snake/snake_game.cpp
@@ -1,10 +1,7 @@
 #include <iostream>
-#include <cmath>
-#include <stdlib.h>
 #include <ncurses.h>
 #include <ctime>
 #include <list>
-#include <sys/ioctl.h>
 
 #include "Game.h"
 
@@ -126,24 +123,6 @@ void print_board() {
   }
 }
 
-void print_prompt() {
-  int halfx = GameVars::MAX_BOARD_X;
-  int halfy = GameVars::MAX_BOARD_Y / 2;
-  for (int i = halfx - 19; i <= halfx + 19; ++i) {
-    for (int j = halfy - 4; j <= halfy + 4; ++j) {
-      mvprintw(j, i, " ");
-    }
-  }
-  for (int i = halfx - 19; i <= halfx + 19; ++i) {
-    mvprintw(halfy - 5, i, "_");
-    mvprintw(halfy + 5, i, "-");
-  }
-  for (int i = halfy - 4; i <= halfy + 4; ++i) {
-    mvprintw(i, halfx - 20, "|");
-    mvprintw(i, halfx + 20, "|");
-  }
-}
-
 bool check_collisions(const std::list<Coords>& snake_list) {
   // Only need to check head
   const Coords head = snake_list.back();
@@ -259,44 +238,4 @@ void set_name(std::list<Coords>& snake_list) {
     it++->symbol = '#';
   }
   it->symbol = 'X';  // head
-}
-
-bool kbhit() {
-    int ch = getch();
-    if (ch != ERR) {
-        ungetch(ch);
-        return true;
-    }
-    else {
-        return false;
-    }
-}
-
-void init_ncurses() {
-  initscr();
-  curs_set(0);
-  cbreak();
-  noecho();
-  nodelay(stdscr, TRUE);  // getch can't block clock
-  keypad(stdscr, TRUE);
-  scrollok(stdscr, TRUE);
-}
-
-bool set_game_size() {
-  struct winsize w;
-  ioctl(0, TIOCGWINSZ, &w);
-
-  int y_save = GameVars::MAX_BOARD_Y;
-  int x_save = GameVars::MAX_BOARD_X;
-
-  GameVars::MAX_BOARD_Y = w.ws_row - 1;
-  GameVars::MAX_BOARD_X = (w.ws_col / 2) - 1;
-
-  if (y_save != GameVars::MAX_BOARD_Y ||
-      x_save != GameVars::MAX_BOARD_X) {
-    // Only resize term if window size changed
-    resizeterm(w.ws_row, w.ws_col);
-    return true;
-  }
-  return false;
 }
